[Resposividade-MobileFirt] refactor: apply atomic design patterns 'This commit refactors the code structure to adhere to atomic design principles. Components are now organized following the atomic design methodology, promoting modularity and maintainability.'
